components/docs_app/requirements.txt
flask>=1.1.0
google-cloud-storage
gunicorn